Working on cleaning up

+++ src/main/java/com/codeup/adlister/controllers/DeleteVGamesServlet.java
@@ -1,7 +1,6 @@
 package com.codeup.adlister.controllers;
 
 import com.codeup.adlister.dao.DaoFactory;
-
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -9,14 +8,20 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+/*
+    Handles deleting of Video Game listings based on MYSQL id assigned to VideoGames
+ */
 @WebServlet(name = "DeleteVGamesServlet", urlPatterns = "/ads/delete")
 public class DeleteVGamesServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-//        request.getPathInfo().substring(1)
-//        request.getParameter("gameID")
-        System.out.println(request.getParameter("gameID"));
         int deleteGame = Integer.parseInt(request.getParameter("gameID"));
+        // Prevent non-logged in users from deleting submissions
+        if (request.getSession().getAttribute("user") == null) {
+            response.sendRedirect("/login");
+            return;
+        }
+        // TODO: Only allow users to delete their own listings
         try {
             DaoFactory.getVgsDao().deleteVGames(deleteGame);
         } catch (Exception e) {
